tests/wpt/web-platform-tests/signed-exchange/sxg-version1b2.tentative.html

<!DOCTYPE html>
<title>Loading SignedHTTPExchange of unsupported version must fail</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="./resources/sxg-util.js"></script>
<body>
<p>Note: This test FAILs if your browser supports Signed Exchange version 1b2.</p>
<script>
promise_test(async (t) => {
  const sxgUrl = get_host_info().HTTPS_ORIGIN +
                 '/signed-exchange/resources/sxg-version1b2.sxg';
  const message = await openSXGInIframeAndWaitForMessage(t, sxgUrl);
  const innerURL = innerURLOrigin() +
                   '/signed-exchange/resources/inner-url.html';
  assert_equals(message.location, innerURL);
  assert_true(message.is_fallback);
}, 'Loading SignedHTTPExchange of unsupported version must fail and fallback redirect.');
</script>
</body>
